[routes]organise message routes
